Разделение Utils на DrawingExtensions и EnumerationExtensions
